add generic tf writer
