+++ ml_tools/irwriter.py
@@ -38,8 +38,6 @@ import cv2
 import random
 import math
 
-crop_rectangle = tools.Rectangle(0, 0, 640 - 1, 480 - 1)
-
 
 def create_tf_example(sample, thermal, filtered, labels):
     """Converts image and annotations to a tf.Example proto.
@@ -90,15 +88,7 @@ def create_tf_example(sample, thermal, filtered, labels):
     image.save(encoded_jpg_io, format="PNG", quality=100, subsampling=0)
     encoded_filtered = encoded_jpg_io.getvalue()
     filtered_key = hashlib.sha256(encoded_filtered).hexdigest()
-    # if frame.mask:
-    # image = Image.fromarray(frame.mask)
-    # image = ImageOps.grayscale(image)
-    # image_id = sample.id
-    #
-    # encoded_jpg_io = io.BytesIO()
-    # image.save(encoded_jpg_io, format="JPEG")
-    # encoded_mask = encoded_jpg_io.getvalue()
-    # mask_key = hashlib.sha256(encoded_mask).hexdigest()
+
     feature_dict = {
         "image/augmented": tfrecord_util.int64_feature(sample.augment),
         "image/height": tfrecord_util.int64_feature(image_height),
@@ -117,103 +107,12 @@ def create_tf_example(sample, thermal, filtered, labels):
         "image/clip_id": tfrecord_util.int64_feature(sample.clip_id),
         "image/track_id": tfrecord_util.int64_feature(sample.track_id),
         "image/filteredencoded": tfrecord_util.bytes_feature(encoded_filtered),
-        # "image/maskkey/sha256": tfrecord_util.bytes_feature(mask_key.encode("utf8")),
-        # "image/maskencoded": tfrecord_util.bytes_feature(encoded_mask),
         "image/format": tfrecord_util.bytes_feature("jpeg".encode("utf8")),
         "image/class/text": tfrecord_util.bytes_feature(sample.label.encode("utf8")),
         "image/class/label": tfrecord_util.int64_feature(labels.index(sample.label)),
-        # "image/object/bbox/xmin": tfrecord_util.float_list_feature([0]),
-        # "image/object/bbox/xmax": tfrecord_util.float_list_feature([1]),
-        # "image/object/bbox/ymin": tfrecord_util.float_list_feature([0]),
-        # "image/object/bbox/ymax": tfrecord_util.float_list_feature([1]),
     }
-    #
-    # num_annotations_skipped = 0
-    # if len(sample.regions) > 0:
-    #     xmin = []
-    #     xmax = []
-    #     ymin = []
-    #     ymax = []
-    #     is_crowd = []
-    #     category_names = []
-    #     category_ids = []
-    #     area = []
-    #     encoded_mask_png = []
-    #     for r, label in zip(sample.regions, sample.labels):
-    #         (x, y, width, height) = r.x, r.y, r.width, r.height
-    #         if width <= 0 or height <= 0:
-    #             num_annotations_skipped += 1
-    #             continue
-    #         if x + width > image_width or y + height > image_height:
-    #             num_annotations_skipped += 1
-    #             continue
-    #         xmin.append(float(x) / image_width)
-    #         xmax.append(float(x + width) / image_width)
-    #         ymin.append(float(y) / image_height)
-    #         ymax.append(float(y + height) / image_height)
-    #         category_id = int(labels.index(label) + 1)
-    #         category_ids.append(category_id)
-    #         category_names.append(label.encode("utf8"))
-    #         area.append(r.area)
-    #     feature_dict.update(
-    #         {
-    #             "image/object/bbox/xmin": tfrecord_util.float_list_feature(xmin),
-    #             "image/object/bbox/xmax": tfrecord_util.float_list_feature(xmax),
-    #             "image/object/bbox/ymin": tfrecord_util.float_list_feature(ymin),
-    #             "image/object/bbox/ymax": tfrecord_util.float_list_feature(ymax),
-    #             "image/object/class/text": tfrecord_util.bytes_list_feature(
-    #                 category_names
-    #             ),
-    #             "image/object/class/label": tfrecord_util.int64_list_feature(
-    #                 category_ids
-    #             ),
-    #             "image/object/area": tfrecord_util.float_list_feature(area),
-    #         }
-    #     )
     example = tf.train.Example(features=tf.train.Features(feature=feature_dict))
     return example
-
-
-def process_job(queue, labels, base_dir, back_thresh):
-    import gc
-
-    pid = os.getpid()
-
-    writer_i = 1
-    name = f"{writer_i}-{pid}.tfrecord"
-
-    options = tf.io.TFRecordOptions(compression_type="GZIP")
-    writer = tf.io.TFRecordWriter(str(base_dir / name), options=options)
-    i = 0
-    saved = 0
-    files = 0
-    while True:
-        i += 1
-        samples = queue.get()
-        try:
-            if samples == "DONE":
-                writer.close()
-                break
-            else:
-                if len(samples) == 0:
-                    continue
-                saved += save_data(samples, writer, labels, back_thresh)
-                files += 1
-                del samples
-                if saved > 10000:
-                    logging.info("Closing old writer")
-                    writer.close()
-                    writer_i += 1
-                    name = f"{writer_i}-{pid}.tfrecord"
-                    logging.info("Opening %s", name)
-                    saved = 0
-                    writer = tf.io.TFRecordWriter(str(base_dir / name), options=options)
-                if i % 100 == 0:
-                    logging.info("Saved %s ", files)
-                    gc.collect()
-                    writer.flush()
-        except:
-            logging.error("Process_job error %s", samples[0].source_file, exc_info=True)
 
 
 def get_data(samples, back_thresh):
@@ -255,8 +154,8 @@ def get_data(samples, back_thresh):
     return data
 
 
-def save_data(samples, writer, labels, back_thresh):
-    sample_data = get_data(samples, back_thresh)
+def save_data(samples, writer, labels, extra_args):
+    sample_data = get_data(samples, extra_args["back_thresh"])
     if sample_data is None:
         return 0
     saved = 0
@@ -275,58 +174,3 @@ def save_data(samples, writer, labels, back_thresh):
             "Could not save data for %s", samples[0].source_file, exc_info=True
         )
     return saved
-
-
-def create_tf_records(
-    dataset, output_path, labels, back_thresh, num_shards=1, cropped=True, augment=False
-):
-    output_path = Path(output_path)
-    if output_path.is_dir():
-        logging.info("Clearing dir %s", output_path)
-        for child in output_path.glob("*"):
-            if child.is_file():
-                child.unlink()
-    output_path.mkdir(parents=True, exist_ok=True)
-    samples_by_source = dataset.get_samples_by_source()
-    source_files = list(samples_by_source.keys())
-    np.random.shuffle(source_files)
-
-    num_labels = len(dataset.labels)
-    logging.info(
-        "writing to output path: %s for %s samples", output_path, len(samples_by_source)
-    )
-    num_processes = 1
-    try:
-        job_queue = Queue()
-        processes = []
-        for i in range(num_processes):
-            p = Process(
-                target=process_job,
-                args=(job_queue, labels, output_path, back_thresh),
-            )
-            processes.append(p)
-            p.start()
-            added = 0
-        for source_file in source_files:
-            job_queue.put((samples_by_source[source_file]))
-            added += 1
-            while job_queue.qsize() > num_processes * 3:
-                logging.info("Sleeping for %s", 10)
-                # give it a change to catch up
-                time.sleep(10)
-
-        logging.info("Processing %d", job_queue.qsize())
-        for i in range(len(processes)):
-            job_queue.put(("DONE"))
-        for process in processes:
-            try:
-                process.join()
-            except KeyboardInterrupt:
-                logging.info("KeyboardInterrupt, terminating.")
-                for process in processes:
-                    process.terminate()
-                exit()
-        logging.info("Saved %s", len(dataset.samples_by_id))
-
-    except:
-        logging.error("Error saving track info", exc_info=True)
